chore: moved the `with` components to `components`
