+++ components/withSideBar.tsx
@@ -1,13 +1,19 @@
+import type { RichTranslationValues } from 'next-intl';
 import type { FC } from 'react';
 
 import SideBar from '@/components/Common/Sidebar';
 import { useSiteNavigation } from '@/hooks/server';
 import type { NavigationKeys } from '@/types';
 
-const WithSideBar: FC<{ keys: NavigationKeys[] }> = ({ keys }) => {
+type WithSideBarProps = {
+  keys: NavigationKeys[];
+  context?: Record<string, RichTranslationValues>;
+};
+
+const WithSideBar: FC<WithSideBarProps> = ({ keys, context }) => {
   const { getSideNavigation } = useSiteNavigation();
 
-  const mappedSidebarItems = getSideNavigation(keys).map(
+  const mappedSidebarItems = getSideNavigation(keys, context).map(
     ([, { label, items }]) => ({
       groupName: label,
       items: items.map(([, { label, link }]) => ({ title: label, url: link })),
