+++ layouts/New/Default.tsx
@@ -1,7 +1,7 @@
 import type { FC, PropsWithChildren } from 'react';
 
-import WithFooter from '@/layouts/New/withFooter';
-import WithNavBar from '@/layouts/New/withNavBar';
+import WithFooter from '@/components/withFooter';
+import WithNavBar from '@/components/withNavBar';
 
 const DefaultLayout: FC<PropsWithChildren> = ({ children }) => (
   <>
